requirements.txt
asgiref==3.3.2
Django~=3.2
django-environ==0.4.5
gunicorn==20.0.4
pytz==2020.1
sqlparse==0.3.1
mysqlclient~=2.1.1

djangorestframework~=3.14.0
djangorestframework-simplejwt~=5.2.2
django-filter~=22.1
PyJWT~=2.6.0
Pillow~=9.3.0
PyYAML==6.0